requirements.txt
gitpython
Faker
requests
pyyaml
coloredlogs